scripts/requirements.txt
PyGithub==2.1.1
python-frontmatter==1.0.0
python-dotenv==1.0.0
